fix : Exception Handling, 알고리즘 수정, 코드 리펙토링 | add : UserInfoAPI

+++ src/main/java/io/devlog/blog/user/service/UserSubServiceImpl.java
@@ -1,6 +1,8 @@
 package io.devlog.blog.user.service;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import io.devlog.blog.config.CustomException;
+import io.devlog.blog.config.enums.Status;
 import io.devlog.blog.user.DTO.SubscribesDTO;
 import io.devlog.blog.user.DTO.UserDTO;
 import io.devlog.blog.user.entity.QSubscribes;
@@ -42,24 +44,33 @@ public class UserSubServiceImpl extends QuerydslRepositorySupport implements Use
     }
 
     @Override
+    public ResponseEntity<?> getUsersSubCount(long userUuid) {
+        long count = sbRepo.findByUserUuid(userUuid).size();
+        if (count == 0) {
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.ok(count);
+        }
+    }
+
+    @Override
     public ResponseEntity<?> getUserSub(long userUuid) {
         try {
             List<Subscribes> finds = sbRepo.findByUserUuid(userUuid);
             if (finds.isEmpty()) {
-                return ResponseEntity.status(404).body("No sub");
+                return ResponseEntity.noContent().build();
             } else {
                 HashMap<Long, UserDTO> subUsers = new HashMap<>();
                 finds.forEach((s) -> {
                     Optional<User> i = userRepo.findByUserUuid(s.getSubUser());
-                    if (i.isEmpty()) {
-                        return;
+                    if (i.isPresent()) {
+                        UserDTO e = UserDTO.toDTO(i.get());
+                        e.setPw(null);
+                        e.setMail(null);
+                        e.setBender(null);
+                        e.setBenderUuid(null);
+                        subUsers.put(i.get().getUserUuId(), e);
                     }
-                    UserDTO e = UserDTO.toDTO(i.get());
-                    e.setPw(null);
-                    e.setMail(null);
-                    e.setBender(null);
-                    e.setBenderUuid(null);
-                    subUsers.put(i.get().getUserUuId(), e);
                 });
                 return ResponseEntity.ok(subUsers);
             }
@@ -69,8 +80,12 @@ public class UserSubServiceImpl extends QuerydslRepositorySupport implements Use
     }
 
     @Override
-    public ResponseEntity<?> addUserSub(SubscribesDTO sbDTO) {
+    public ResponseEntity<?> addUserSub(long userUuid, SubscribesDTO sbDTO) {
         try {
+            sbDTO.setUserUuid(userUuid);
+            if (sbRepo.existsBySubUserUuid(sbDTO.getSubUser())) {
+                throw new CustomException(Status.CONFLICT);
+            }
             return ResponseEntity.ok(sbRepo.save(sbDTO.toEntity()));
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e);
@@ -80,6 +95,9 @@ public class UserSubServiceImpl extends QuerydslRepositorySupport implements Use
     @Override
     public ResponseEntity<?> deleteUserSub(long subUuid) {
         try {
+            if (!sbRepo.existsBySubUserUuid(subUuid)) {
+                return ResponseEntity.badRequest().body(new CustomException(Status.NO_CONTENT));
+            }
             sbRepo.deleteById(subUuid);
             return ResponseEntity.ok("Deleted");
         } catch (Exception e) {
